+++ Assets/Scripts/JointStatePublisher.cs
@@ -52,7 +52,7 @@ public class JointStatePublisher : MonoBehaviour
             uint secs = (uint)sim_time;
             uint nsecs = (uint)((sim_time % 1) * 1e9);
             message.header.frame_id = "world";
-            message.header.stamp.sec = secs;
+            message.header.stamp.sec = (int)secs;
             message.header.stamp.nanosec = nsecs;
             message.position = new double[joints.Count];
             message.velocity = new double[joints.Count];
