+++ Assets/Scripts/WorldClock.cs
@@ -25,7 +25,7 @@ public class WorldClock : MonoBehaviour
         float sim_time = Time.time;
         uint secs = (uint)sim_time;
         uint nsecs = (uint)((sim_time % 1) * 1e9);
-        message.clock.sec = secs;
+        message.clock.sec = (int)secs;
         message.clock.nanosec = nsecs;
         ros.Send(topicName, message);
     }
